move the logic responsible for getting executable version from RootCommand to VersionOption:
- make it clear when Reflection is used
- don't load and compile Assembly type when RootCommand is used for the first time
